# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Biz Connect Academy
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LeadInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  city: string;
}

export interface Lead {
  id: number;
  name: string;
  phone: string;
  city: string;
  createdAt: string;
}

export interface LeadStatDay {
  date: string;
  count: number;
}

export interface LeadsStats {
  total: number;
  byDay: LeadStatDay[];
}

export interface AdminCredentials {
  password: string;
}

export interface AdminVerifyResult {
  valid: boolean;
}

export interface SiteContent {
  id: number;
  heroTitle: string;
  heroSubtitle: string;
  heroCtaText: string;
  memberCount: string;
  memberCountLabel: string;
  geoAvailability: string;
  videoUrl: string;
  offerPrice: string;
  offerOriginalPrice: string;
  offerLabel: string;
  level1Name: string;
  level1Amount: string;
  level2Name: string;
  level2Amount: string;
  level3Name: string;
  level3Amount: string;
  whatsappNumber: string;
  heroImageUrl?: string;
  signupUrl?: string;
  communityImageUrl?: string;
  countriesIconUrl?: string;
  telegramLink?: string;
  supportPhone1?: string;
  supportPhone2?: string;
  updatedAt: string;
}

export interface SiteContentInput {
  heroTitle?: string;
  heroSubtitle?: string;
  heroCtaText?: string;
  memberCount?: string;
  memberCountLabel?: string;
  geoAvailability?: string;
  videoUrl?: string;
  offerPrice?: string;
  offerOriginalPrice?: string;
  offerLabel?: string;
  level1Name?: string;
  level1Amount?: string;
  level2Name?: string;
  level2Amount?: string;
  level3Name?: string;
  level3Amount?: string;
  whatsappNumber?: string;
  heroImageUrl?: string;
  signupUrl?: string;
  communityImageUrl?: string;
  countriesIconUrl?: string;
  telegramLink?: string;
  supportPhone1?: string;
  supportPhone2?: string;
}

export interface DeleteResult {
  success: boolean;
}

export interface Training {
  id: number;
  title: string;
  bannerUrl: string;
  linkUrl: string;
  sortOrder: number;
}

export interface TrainingInput {
  /** @minLength 1 */
  title: string;
  bannerUrl?: string;
  linkUrl?: string;
  sortOrder?: number;
}

export interface Testimonial {
  id: number;
  name: string;
  country: string;
  duration: string;
  text: string;
  mediaUrl: string;
  mediaType: string;
  sortOrder: number;
}

export interface TestimonialInput {
  /** @minLength 1 */
  name: string;
  country?: string;
  duration?: string;
  text?: string;
  mediaUrl?: string;
  mediaType?: string;
  sortOrder?: number;
}

export interface PortfolioItem {
  id: number;
  imageUrl: string;
  caption: string;
  sortOrder: number;
}

export interface PortfolioItemInput {
  /** @minLength 1 */
  imageUrl: string;
  caption?: string;
  sortOrder?: number;
}

export interface Partner {
  id: number;
  name: string;
  logoUrl: string;
  sortOrder: number;
}

export interface PartnerInput {
  name?: string;
  /** @minLength 1 */
  logoUrl: string;
  sortOrder?: number;
}

export interface PaymentMethod {
  id: number;
  name: string;
  logoUrl: string;
  sortOrder: number;
}

export interface PaymentMethodInput {
  name?: string;
  /** @minLength 1 */
  logoUrl: string;
  sortOrder?: number;
}

export interface UploadUrlRequest {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResult {
  uploadURL: string;
  objectPath: string;
}

export interface Service {
  id: number;
  title: string;
  description: string;
  iconUrl: string;
  sortOrder: number;
}

export interface ServiceInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  iconUrl?: string;
  sortOrder?: number;
}

export interface FeatureItem {
  id: number;
  section: string;
  label: string;
  sortOrder: number;
}

export interface FeatureItemInput {
  section?: string;
  /** @minLength 1 */
  label: string;
  sortOrder?: number;
}

export interface HelpVideo {
  id: number;
  title: string;
  description: string;
  videoUrl: string;
  sortOrder: number;
}

export interface HelpVideoInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  /** @minLength 1 */
  videoUrl: string;
  sortOrder?: number;
}

export interface Notification {
  id: number;
  name: string;
  city: string;
  action: string;
  time: string;
}