# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Biz Connect Academy
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead (public CTA form submission)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads (admin only)
      parameters:
        - name: x-admin-password
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leads/export:
    get:
      operationId: exportLeads
      tags: [leads]
      summary: Export leads as CSV (admin only)
      parameters:
        - name: x-admin-password
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leads/stats:
    get:
      operationId: getLeadsStats
      tags: [leads]
      summary: Get lead statistics per day (admin only)
      parameters:
        - name: x-admin-password
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stats grouped by day
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadsStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/verify:
    post:
      operationId: verifyAdmin
      tags: [admin]
      summary: Verify admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Password valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminVerifyResult"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    LeadInput:
      type: object
      required: [name, phone, city]
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
    Lead:
      type: object
      required: [id, name, phone, city, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        phone:
          type: string
        city:
          type: string
        createdAt:
          type: string
    LeadStatDay:
      type: object
      required: [date, count]
      properties:
        date:
          type: string
        count:
          type: number
    LeadsStats:
      type: object
      required: [total, byDay]
      properties:
        total:
          type: number
        byDay:
          type: array
          items:
            $ref: "#/components/schemas/LeadStatDay"
    AdminCredentials:
      type: object
      required: [password]
      properties:
        password:
          type: string
    AdminVerifyResult:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
...