# lib/api-zod/src/generated/types/adminVerifyResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Biz Connect Academy
 * OpenAPI spec version: 0.1.0
 */

export interface AdminVerifyResult {
  valid: boolean;
}