# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Biz Connect Academy
 * OpenAPI spec version: 0.1.0
 */

export interface LeadInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  city: string;
}